6-deployed-agents/knowledge-base/image-generator-agent/pyproject.toml
[tool.poetry]
name = "image-generator-agent"
version = "0.1.0"
description = "This agent is integrated with LangChain's DallEAPIWrapper to generate images based on textual descriptions."
authors = ["Alejandro <alejandro.madrigal@fetch.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
uagents = "^0.17.0"
langchain-community = "^0.0.25"
langchain-openai = ">=0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"